fix(docker): properly pass control down to processes
Without execing the process, the process will not be able to receive signals
such as SIGINT from the parent process and instead the shell scripts receive the signal.
This is important for the gunicorn/capture/indiserver processes to be able to receive
signals to terminate gracefully, otherwise the shell script will exit leaving the gunicorn/capture/indiserver
process, then eventually the timeout will be reached and the container will be killed forcefully.